requirements.txt
APScheduler==3.11.0
cloudinary==1.44.0
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.4
pytest==8.3.5
python-dotenv==1.1.0
python_jose==3.4.0
SQLAlchemy==2.0.40
